requirements.txt
python=3.9
pytorch-cuda=12.1
pytorch=2.1.1
torchvision=0.16.1
torchaudio=2.1.1
yacs=0.1.8
cython=3.0.5